Screen edited with mock data

+++ app/src/main/java/com/aayar94/weather/data/model/weather/Current.kt
@@ -10,7 +10,7 @@ data class Current(
     val sunrise: Int,
     val sunset: Int,
     val temp: Double,
-    val uvi: Double,
+    val uvi: Int,
     val visibility: Int,
     val weather: List<Weather>,
     val wind_deg: Int,
